back-refac: reorganizando os pacotes e nomeclaturas, criando pacote repetir transação
