+++ back/MeuBolso/src/main/java/com/projetointegrado/MeuBolso/config/InjecaoDependeciaConfig.java
@@ -12,8 +12,8 @@ import com.projetointegrado.MeuBolso.tipoConta.ITipoContaService;
 import com.projetointegrado.MeuBolso.tipoConta.TipoContaService;
 import com.projetointegrado.MeuBolso.transacao.ITransacaoService;
 import com.projetointegrado.MeuBolso.transacao.TransacaoService;
-import com.projetointegrado.MeuBolso.transacaoFixa.ITransacaoFixaService;
-import com.projetointegrado.MeuBolso.transacaoFixa.TransacaoFixaService;
+import com.projetointegrado.MeuBolso.transacaoRecorrente.ITransacaoFixaService;
+import com.projetointegrado.MeuBolso.transacaoRecorrente.TransacaoFixaService;
 import com.projetointegrado.MeuBolso.usuario.IUsuarioService;
 import com.projetointegrado.MeuBolso.usuario.UsuarioService;
 import org.springframework.context.annotation.Bean;
