+++ back/MeuBolso/src/main/java/com/projetointegrado/MeuBolso/repetirTransacao/TransacaoRepeticaoExecutor.java
@@ -1,4 +1,4 @@
-package com.projetointegrado.MeuBolso.transacaoFixa;
+package com.projetointegrado.MeuBolso.repetirTransacao;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
