+++ back/MeuBolso/src/main/java/com/projetointegrado/MeuBolso/transacao/TransacaoService.java
@@ -8,7 +8,7 @@ import com.projetointegrado.MeuBolso.globalExceptions.AcessoNegadoException;
 import com.projetointegrado.MeuBolso.globalExceptions.EntidadeNaoEncontradaException;
 import com.projetointegrado.MeuBolso.transacao.dto.TransacaoSaveDTO;
 import com.projetointegrado.MeuBolso.transacao.dto.TransacaoDTO;
-import com.projetointegrado.MeuBolso.transacaoFixa.TransacaoRepeticaoService;
+import com.projetointegrado.MeuBolso.repetirTransacao.TransacaoRepeticaoService;
 import com.projetointegrado.MeuBolso.usuario.Usuario;
 import com.projetointegrado.MeuBolso.usuario.UsuarioValidateService;
 import org.springframework.beans.factory.annotation.Autowired;
