+++ back/MeuBolso/src/main/java/com/projetointegrado/MeuBolso/transacaoRecorrente/TransacaoFixaRepository.java
@@ -1,4 +1,4 @@
-package com.projetointegrado.MeuBolso.transacaoFixa;
+package com.projetointegrado.MeuBolso.transacaoRecorrente;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
