+++ back/MeuBolso/src/main/java/com/projetointegrado/MeuBolso/transacaoRecorrente/dto/TransacaoFixaSaveDTO.java
@@ -1,8 +1,8 @@
-package com.projetointegrado.MeuBolso.transacaoFixa.dto;
+package com.projetointegrado.MeuBolso.transacaoRecorrente.dto;
 
 import com.projetointegrado.MeuBolso.globalConstraints.validEnum.ValidEnum;
 import com.projetointegrado.MeuBolso.transacao.TipoTransacao;
-import com.projetointegrado.MeuBolso.transacaoFixa.Periodicidade;
+import com.projetointegrado.MeuBolso.transacaoRecorrente.Periodicidade;
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
